Wired up WebSockets to work in conjunction with HTML according to the getting started tutorial of Channels for Django. Now the client can send a WebSocket message to the consumer and the consumer will echo the message back to the client.
In the context of my app, the client can choose an opponent to challenge from a list of opponents, challenge the opponent, and the Player (User) id of the opponent will get added to the list of opponents for the client. This in itself isnt useful, but it will be once I enable a Channel layer that allows client WebSocket instances to communicate.